+++ apps/remix-ide-e2e/src/tests/proxy_oz_v5.test.ts
@@ -16,9 +16,6 @@ module.exports = {
     return sources
   },
 
-  'Should set the compiler version to 8.19': function(browser: NightwatchBrowser) {
-    browser.setSolidityCompilerVersion('soljson-v0.8.19+commit.7dd6d404.js')
-  },
 
   'Should show deploy proxy option for UUPS upgradeable contract #group1': function (browser: NightwatchBrowser) {
     browser
@@ -29,7 +26,7 @@ module.exports = {
       .click('[data-id="compilerContainerCompileBtn"]')
       .clickLaunchIcon('filePanel')
       .isVisible({
-        selector: '*[data-id="treeViewDivtreeViewItem.deps/npm/@openzeppelin/contracts-upgradeable@4.8.3/proxy/beacon/IBeaconUpgradeable.sol"]',
+        selector: '*[data-id="treeViewDivtreeViewItem.deps/npm/@openzeppelin/contracts-upgradeable/utils/ContextUpgradeable.sol"]',
         timeout: 120000,
         suppressNotFoundErrors: true
       })
@@ -37,7 +34,7 @@ module.exports = {
       .click('[data-id="compilerContainerCompileBtn"]')
       .clickLaunchIcon('filePanel')
       .isVisible({
-        selector: '*[data-id="treeViewDivtreeViewItem.deps/npm/@openzeppelin/contracts-upgradeable@4.8.3/proxy/beacon/IBeaconUpgradeable.sol"]',
+        selector: '*[data-id="treeViewDivtreeViewItem.deps/npm/@openzeppelin/contracts-upgradeable/utils/ContextUpgradeable.sol"]',
         timeout: 120000,
         suppressNotFoundErrors: true
       })
@@ -45,7 +42,7 @@ module.exports = {
       .click('[data-id="compilerContainerCompileBtn"]')
       .clickLaunchIcon('filePanel')
       .waitForElementVisible({
-        selector: '*[data-id="treeViewDivtreeViewItem.deps/npm/@openzeppelin/contracts-upgradeable@4.8.3/proxy/beacon/IBeaconUpgradeable.sol"]',
+        selector: '*[data-id="treeViewDivtreeViewItem.deps/npm/@openzeppelin/contracts-upgradeable/utils/ContextUpgradeable.sol"]',
         timeout: 120000,
       })
       .clickLaunchIcon('solidity')
@@ -84,6 +81,7 @@ module.exports = {
       .verify.visible('[data-id="contractGUIDeployWithProxyLabel"]')
       .waitForElementPresent('[data-id="contractGUIDeployWithProxyLabel"]')
       .click('[data-id="contractGUIDeployWithProxyLabel"]')
+      .setValue('[data-id="initializeInputs-initialOwner"]', '0x5B38Da6a701c568545dCfcB03FcB875f56beddC4')
       .createContract('')
       .waitForElementContainsText('[data-id="udappNotifyModalDialogModalTitle-react"]', 'Deploy Implementation & Proxy (ERC1967)')
       .waitForElementVisible('[data-id="udappNotify-modal-footer-ok-react"]')
@@ -93,7 +91,7 @@ module.exports = {
       .click('[data-id="confirmProxyDeployment-modal-footer-ok-react"]')
       .waitForElementPresent('[data-id="universalDappUiTitleExpander0"]')
       .waitForElementPresent('[data-id="universalDappUiTitleExpander1"]')
-      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Deploying ERC1967 < 5.0.0 as proxy...', 60000)
+      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Deploying ERC1967 >= 5.0.0 as proxy...')
   },
 
   'Should interact with deployed contract via ERC1967 (proxy) #group1': function (browser: NightwatchBrowser) {
@@ -135,6 +133,7 @@ module.exports = {
       .setValue('//*[@id="runTabView"]/div/div[2]/div[3]/div[1]/div/div[1]/div[4]/div/div[1]/input', 'Remix')
       .setValue('//*[@id="runTabView"]/div/div[2]/div[3]/div[1]/div/div[1]/div[4]/div/div[2]/input', "R")
       .useCss()
+      .setValue('[data-id="initializeInputs-initialOwner"]', '0x5B38Da6a701c568545dCfcB03FcB875f56beddC4')
       .createContract('')
       .waitForElementContainsText('[data-id="udappNotifyModalDialogModalTitle-react"]', 'Deploy Implementation & Proxy (ERC1967)')
       .waitForElementVisible('[data-id="udappNotify-modal-footer-ok-react"]')
@@ -144,7 +143,7 @@ module.exports = {
       .click('[data-id="confirmProxyDeployment-modal-footer-ok-react"]')
       .waitForElementPresent('[data-id="universalDappUiTitleExpander0"]')
       .waitForElementPresent('[data-id="universalDappUiTitleExpander1"]')
-      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Deploying ERC1967 < 5.0.0 as proxy...', 60000)
+      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Deploying ERC1967 >= 5.0.0 as proxy...')
   },
 
   'Should interact with initialized contract to verify parameters #group1': function (browser: NightwatchBrowser) {
@@ -186,6 +185,7 @@ module.exports = {
       .waitForElementVisible('[data-id="proxy-dropdown-items"]')
       .assert.textContains('[data-id="proxy-dropdown-items"]', shortenedFirstAddress)
       .assert.textContains('[data-id="proxy-dropdown-items"]', shortenedLastAddress)
+
       .click('[data-id="proxyAddress1"]')
       .createContract('')
       .waitForElementContainsText('[data-id="udappNotifyModalDialogModalTitle-react"]', 'Deploy Implementation & Update Proxy')
@@ -199,7 +199,7 @@ module.exports = {
         })
       .waitForElementPresent('[data-id="universalDappUiTitleExpander0"]')
       .waitForElementPresent('[data-id="universalDappUiTitleExpander1"]')
-      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Using ERC1967 < 5.0.0 for the proxy upgrade..', 60000)
+      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Using ERC1967 >= 5.0.0 for the proxy upgrade...')
   },
 
   'Should interact with upgraded function in contract MyTokenV2 #group1': function (browser: NightwatchBrowser) {
@@ -238,18 +238,30 @@ module.exports = {
       .click('[data-id="confirmProxyDeployment-modal-footer-ok-react"]')
       .waitForElementPresent('[data-id="universalDappUiTitleExpander0"]')
       .waitForElementPresent('[data-id="universalDappUiTitleExpander1"]')
-      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Using ERC1967 < 5.0.0 for the proxy upgrade..', 60000)
+      .waitForElementContainsText('*[data-id="terminalJournal"]', 'Using ERC1967 >= 5.0.0 for the proxy upgrade...')
   },
 
   'Should interact with upgraded contract through provided proxy address #group1': function (browser: NightwatchBrowser) {
     browser
+      .clearConsole()
       .clickInstance(1)
       .perform((done) => {
         browser.testConstantFunction(firstProxyAddress, 'version - call', null, '0:\nstring: MyTokenV2!').perform(() => {
           done()
         })
       })
-      .end()
+  },
+  'Should debug the call': function(browser: NightwatchBrowser) {
+    browser
+    .debugTransaction(0)
+    .waitForElementVisible({
+      locateStrategy: 'xpath',
+      selector: '//*[@data-id="treeViewLivm trace step" and contains(.,"7")]',
+      timeout: 60000
+    })
+    .goToVMTraceStep(129)
+    .waitForElementContainsText('*[data-id="functionPanel"]', 'version()', 60000)
+    .end()
   }
 }
 
@@ -259,12 +271,12 @@ const sources = [
     'myTokenV1.sol': {
       content: `
       // SPDX-License-Identifier: MIT
-      pragma solidity ^0.8.4;
+      pragma solidity ^0.8.20;
       
-      import "@openzeppelin/contracts-upgradeable@4.8.3/token/ERC721/ERC721Upgradeable.sol";
-      import "@openzeppelin/contracts-upgradeable@4.8.3/access/OwnableUpgradeable.sol";
-      import "@openzeppelin/contracts-upgradeable@4.8.3/proxy/utils/Initializable.sol";
-      import "@openzeppelin/contracts-upgradeable@4.8.3/proxy/utils/UUPSUpgradeable.sol";
+      import "@openzeppelin/contracts-upgradeable/token/ERC721/ERC721Upgradeable.sol";
+      import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
+      import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
+      import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
       
       contract MyToken is Initializable, ERC721Upgradeable, OwnableUpgradeable, UUPSUpgradeable {
           /// @custom:oz-upgrades-unsafe-allow constructor
@@ -272,9 +284,9 @@ const sources = [
               _disableInitializers();
           }
       
-          function initialize() initializer public {
+          function initialize(address initialOwner) initializer public {
               __ERC721_init("MyToken", "MTK");
-              __Ownable_init();
+              __Ownable_init(initialOwner);
               __UUPSUpgradeable_init();
           }
       
@@ -289,6 +301,8 @@ const sources = [
   }, {
     'myTokenV2.sol': {
       content: `
+      // SPDX-License-Identifier: MIT
+      pragma solidity ^0.8.20;
       import "./myTokenV1.sol";
 
       contract MyTokenV2 is MyToken {
@@ -302,12 +316,12 @@ const sources = [
     'initializeProxy.sol': {
       content: `
       // SPDX-License-Identifier: MIT
-      pragma solidity ^0.8.4;
+      pragma solidity ^0.8.20;
       
-      import "@openzeppelin/contracts-upgradeable@4.8.3/token/ERC721/ERC721Upgradeable.sol";
-      import "@openzeppelin/contracts-upgradeable@4.8.3/access/OwnableUpgradeable.sol";
-      import "@openzeppelin/contracts-upgradeable@4.8.3/proxy/utils/Initializable.sol";
-      import "@openzeppelin/contracts-upgradeable@4.8.3/proxy/utils/UUPSUpgradeable.sol";
+      import "@openzeppelin/contracts-upgradeable/token/ERC721/ERC721Upgradeable.sol";
+      import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
+      import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
+      import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
       
       contract MyInitializedToken is Initializable, ERC721Upgradeable, OwnableUpgradeable, UUPSUpgradeable {
           /// @custom:oz-upgrades-unsafe-allow constructor
@@ -315,9 +329,9 @@ const sources = [
               _disableInitializers();
           }
       
-          function initialize(string memory tokenName, string memory tokenSymbol) initializer public {
+          function initialize(string memory tokenName, string memory tokenSymbol, address initialOwner) initializer public {
               __ERC721_init(tokenName, tokenSymbol);
-              __Ownable_init();
+              __Ownable_init(initialOwner);
               __UUPSUpgradeable_init();
           }
